- get rid of TempStorageKeys enum - create configurations on save, if needed - don't show fields if you clear template - default folder to root
