--- src/Utils/Enums.ts
@@ -1,16 +1,0 @@
-export enum TempStorageKeys {
-  Filename = 'Filename',
-  Documentname = 'Documentname',
-  DocTemplate = 'DocTemplate',
-  Action = 'Action',
-  Fileurl = 'Fileurl',
-  Destinationfolder = 'Destinationfolder',
-  Fileextension = 'Fileextension',
-  ContextPageAbsoluteUrl = 'ContextPageAbsoluteUrl',
-  PageOrigin = 'PageOrigin',
-  Maping = 'Maping',
-  Filecontenttype = 'Filecontenttype',
-  LContType = 'LContType',
-  configname = 'configname',
-  Filemetadata = 'Filemetadata',
-}
